Fixed #580

+++ firmware/include/functions/codeplug.h
@@ -79,7 +79,7 @@ typedef struct struct_codeplugRxGroup
 	char name[16];
 	uint16_t contacts[32];
 	int	NOT_IN_CODEPLUG_numTGsInGroup;// NOT IN THE
-	uint16_t NOT_IN_CODEPLUG_contactsTG[32];
+	uint32_t NOT_IN_CODEPLUG_contactsTG[32];
 } struct_codeplugRxGroup_t;
 
 typedef struct struct_codeplugContact
